projects/homebase-main-litlabsai/github/_archive_all_apps/agent-zero/requirements.txt
openai
anthropic
python-dotenv
requests
termcolor
google-generativeai
fastapi
uvicorn[standard]
python-multipart
pydantic
websockets
